Customize installation process with prompts

+++ scripts/install.sh
@@ -20,11 +20,13 @@ fi
 gecho 'Linking files'
 source $DIR/scripts/bootstrap.sh
 
-gecho 'Generating SSH key'
-ssh-keygen -t rsa -b 4096 -C $EMAIL -f ~/.ssh/github.key
-eval $(ssh-agent -s)
-ssh-add ~/.ssh/github.key
-git --git-dir $DIR/.git remote set-url origin git@github.com:$REPO
+if prompt 'Generate SSH key?'; then
+    gecho 'Generating SSH key'
+    ssh-keygen -t rsa -b 4096 -C $EMAIL -f ~/.ssh/github.key
+    eval $(ssh-agent -s)
+    ssh-add ~/.ssh/github.key
+    git --git-dir $DIR/.git remote set-url origin git@github.com:$REPO
+fi
 
 if is_windows; then
     gecho 'Installing direnv'
@@ -34,30 +36,44 @@ if is_windows; then
     chmod +x $DIRENV
 fi
 
-gecho 'Installing fisherman'
-FISHDIR=~/.config/fish
-curl -Lo $FISHDIR/functions/fisher.fish --create-dirs https://git.io/fisher
+if prompt 'Install fisherman and fish plugins?'; then
+    gecho 'Installing fisherman'
+    FISHDIR=~/.config/fish
+    curl -Lo $FISHDIR/functions/fisher.fish --create-dirs https://git.io/fisher
 
-gecho 'Installing fish plugins'
-fish -c 'fisher'
-echo 'eval (pipenv --completion)' > $FISHDIR/completions/pipenv.fish
-
-gecho 'Upgrading pip'
-if is_windows; then
-    python -m pip install --upgrade pip
-else
-    pip install --upgrade pip
+    gecho 'Installing fish plugins'
+    fish -c 'fisher'
+    echo 'eval (pipenv --completion)' > $FISHDIR/completions/pipenv.fish
 fi
 
-gecho 'Installing Python packages'
-pip install -r $DIR/support/requirements.txt
-curl -sSL https://raw.githubusercontent.com/sdispater/poetry/master/get-poetry.py | python
-poetry completions fish > ~/.config/fish/completions/poetry.fish
-poetry config settings.virtualenvs.in-project true
+if prompt 'Upgrade pip?'; then
+    gecho 'Upgrading pip'
+    if is_windows; then
+        python -m pip install --upgrade pip
+    else
+        pip install --upgrade pip
+    fi
+fi
 
-gecho 'Installing Powerline fonts'
-git clone https://github.com/powerline/fonts.git --depth=1
-cd fonts
-./install.sh
-cd ..
-rm -rf fonts
+if is_windows; then
+    if prompt 'Install Python packages?'; then
+        gecho 'Installing Python packages'
+        pip install -r $DIR/support/requirements.txt
+    fi
+fi
+
+if prompt 'Install poetry?'; then
+    gecho 'Installing poetry'
+    curl -sSL https://raw.githubusercontent.com/sdispater/poetry/master/get-poetry.py | python
+    poetry completions fish > ~/.config/fish/completions/poetry.fish
+    poetry config settings.virtualenvs.in-project true
+fi
+
+if prompt 'Install Powerline fonts?'; then
+    gecho 'Installing Powerline fonts'
+    git clone https://github.com/powerline/fonts.git --depth=1
+    cd fonts
+    ./install.sh
+    cd ..
+    rm -rf fonts
+fi
